+++ mkdocs-website/docs/en/development/changes.md
@@ -164,7 +164,7 @@ Example:
 			2: 4021313248,
 		},
 		Assets: application.AssetOptions{
-			FS: assets,
+			Handler: application.AssetFileServerFS(assets),
 		},
 		Mac: application.MacOptions{
 			ApplicationShouldTerminateAfterLastWindowClosed: true,